requirements.txt
streamlit
openai
pandas
openpyxl
pdfplumber
PyPDF2
python-dotenv
google-api-python-client
google-auth
google-auth-httplib2
google-auth-oauthlib